caching: create a PersistenceService in charge of caching logic
