+++ Weather-DVT/Scenes/Weather/Weather-ViewModel.swift
@@ -32,13 +32,17 @@ extension WeatherView {
         }
 
         private var locationManager: LocationManager?
-        private var modelContext: ModelContext?
+        private var persistenceService: PersistenceService?
         private let service: WeatherService
         private var cancellables = Set<AnyCancellable>()
 
-        init(locationManager: LocationManager, modelContext: ModelContext, service: WeatherService = DefaultWeatherService()) {
+        init(
+            locationManager: LocationManager,
+            persistenceService: PersistenceService,
+            service: WeatherService = DefaultWeatherService()
+        ) {
             self.locationManager = locationManager
-            self.modelContext = modelContext
+            self.persistenceService = persistenceService
             self.service = service
 
             bindLocation()
@@ -106,7 +110,8 @@ extension WeatherView {
             errorMessage = nil
 
             // check cache based on location kind
-            if let cachedLocation = loadCachedWeather(for: location), let cachedWeather = cachedLocation.weather {
+            if let cachedLocation = persistenceService?.fetchCachedWeather(for: location),
+               let cachedWeather = cachedLocation.weather {
                 self.weather = Weather(from: cachedWeather)
                 self.dailySummaries = cachedLocation.forecast
                     .map { Weather(from: $0) }
@@ -131,7 +136,7 @@ extension WeatherView {
             switch current {
             case .success(let weather):
                 self.weather = weather
-                updateCache(for: location, with: weather)
+                persistenceService?.updateCache(for: location, with: weather)
             case .failure(let error):
                 errorMessage = error.localizedDescription
             }
@@ -139,7 +144,7 @@ extension WeatherView {
             switch forecast {
             case .success(let forecast):
                 getDailySummaries(from: forecast)
-                updateCache(for: location, with: dailySummaries)
+                persistenceService?.updateCache(for: location, with: dailySummaries)
             case .failure(let error):
                 errorMessage = error.localizedDescription
             }
@@ -171,94 +176,6 @@ extension WeatherView {
                     maxTempInCelcius: maxTemp
                 )
             }.sorted(by: { $0.date < $1.date })
-        }
-
-        private func loadCachedWeather(for location: WeatherLocation) -> CachedLocation? {
-            switch location.kind {
-            case .temporary:
-                // temporary locations (from search results) never have cached data
-                return nil
-            case .saved(let id):
-                // for a favourite, find it by its ID
-                let descriptor = FetchDescriptor<CachedLocation>(predicate: #Predicate { $0.id == id })
-                return try? modelContext?.fetch(descriptor).first
-            case .current:
-                // for current location, find it using the unique flag
-                let descriptor = FetchDescriptor<CachedLocation>(predicate: #Predicate { $0.isCurrentUserLocation })
-                return try? modelContext?.fetch(descriptor).first
-            }
-        }
-
-        private func updateCache(for location: WeatherLocation, with freshWeather: Weather) {
-            guard let locationToUpdate = findOrCreateLocation(for: location) else { return }
-
-            // update or create a new CachedCurrentWeather
-            if let cache = locationToUpdate.weather {
-                cache.currentTempCelcius = freshWeather.currentTempInCelcius
-                cache.minTempCelcius = freshWeather.minTempInCelcius
-                cache.maxTempCelcius = freshWeather.maxTempInCelcius
-                cache.main = freshWeather.main
-                cache.lastUpdated = .now
-            } else {
-                let newCache = CachedCurrentWeather(
-                    currentTempCelcius: freshWeather.currentTempInCelcius,
-                    minTempCelcius: freshWeather.minTempInCelcius,
-                    maxTempCelcius: freshWeather.maxTempInCelcius,
-                    main: freshWeather.main
-                )
-                locationToUpdate.weather = newCache
-            }
-
-            try? modelContext?.save()
-        }
-
-        private func updateCache(for location: WeatherLocation, with summaries: [Weather]) {
-            guard let locationToUpdate = findOrCreateLocation(for: location) else { return }
-
-            // delete old forecasts to ensure no stale data
-            locationToUpdate.forecast.removeAll()
-
-            // create new CachedForecastWeather
-            for summary in summaries {
-                let forecast = CachedForecastWeather(
-                    currentTempCelcius: summary.currentTempInCelcius,
-                    minTempCelcius: summary.minTempInCelcius,
-                    maxTempCelcius: summary.maxTempInCelcius,
-                    main: summary.main,
-                    date: summary.date
-                )
-                forecast.location = locationToUpdate
-                locationToUpdate.forecast.append(forecast)
-            }
-        }
-
-        private func findOrCreateLocation(for location: WeatherLocation) -> CachedLocation? {
-            switch location.kind {
-            case .temporary:
-                // never save weather for search results
-                return nil
-            case .saved(let id):
-                // use ID
-                let descriptor = FetchDescriptor<CachedLocation>(predicate: #Predicate { $0.id == id })
-                return try? modelContext?.fetch(descriptor).first
-            case .current:
-                // use unique flag
-                let descriptor = FetchDescriptor<CachedLocation>(predicate: #Predicate { $0.isCurrentUserLocation })
-                if let currentLocation = try? modelContext?.fetch(descriptor).first {
-                    return currentLocation
-                } else {
-                    // create a new location
-                    let newLocation = CachedLocation(
-                        name: "My Location", // TODO: reverse geocode
-                        region: "",
-                        latitude: location.latitude,
-                        longitude: location.longitude,
-                        isCurrentUserLocation: true
-                    )
-                    modelContext?.insert(newLocation)
-                    return newLocation
-                }
-            }
         }
     }
 }
